feat: SOLID principles - Open/Closed principle
